refactor: update imports to use shadcn directory in project-related pages
