+++ app/users/dashboard/projects/[id]/edit/page.tsx
@@ -12,8 +12,8 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
 import { Save, Loader2 } from 'lucide-react';
 import Link from 'next/link';
 import { ProjectStatus } from '@/types/project/project-status';
